refactored settings tabs components
