+++ app/src/Settings.tsx
@@ -4,7 +4,7 @@ import { observer } from "mobx-react-lite";
 import { Column } from "./SettingsTabs/Column";
 import { Time } from "./SettingsTabs/Time";
 import { Font } from "./SettingsTabs/Font";
-import { MapPoints } from "./SettingsTabs/MapPoints";
+import { MapPoints } from "./SettingsTabs/map_points/MapPoints";
 import { Datapacks } from "./SettingsTabs/Datapack";
 import { useTheme } from "@mui/material/styles";
 import { TSCTabs, TSCTab } from "./components";
