+++ app/src/SettingsTabs/map_points/Legend.css
